refactor: move toast component to @/components/Toasts.svelte
